Update New-CsHybridTelephoneNumber.md
"This PowerShell command is being deprecated and will not be available after April 7, 2025. Teams administrators will need to use the new PowerShell command --> "New-CsOnlineDirectRoutingTelephoneNumberUploadOrder" to add a telephone number to the tenant that can be used for Audio Conferencing with Direct Routing for GCC High and DoD clouds. Detailed instructions on how to use the new cmdlets can be found here --> New-CsOnlineDirectRoutingTelephoneNumberUploadOrder (MicrosoftTeamsPowerShell) | Microsoft Learn